refactor: extract payjoin module and fix platform send-flow observers
