+++ android/app/src/main/java/org/bitcoinppl/cove/WalletManager.kt
@@ -79,12 +79,9 @@ class WalletManager :
     var errorAlert by mutableStateOf<WalletErrorAlert?>(null)
     var sendFlowErrorAlert by mutableStateOf<TaggedItem<SendFlowErrorAlert>?>(null)
 
-    // set to true when a payjoin transaction has been broadcast (success or fallback)
-    var payjoinTxBroadcast by mutableStateOf(false)
-
-    fun resetPayjoinTxBroadcast() {
-        payjoinTxBroadcast = false
-    }
+    // non-null when a payjoin transaction has been broadcast (success or fallback);
+    // TaggedItem ensures a new unique key each time so Compose always re-fires the observer
+    var payjoinTxBroadcast by mutableStateOf<TaggedItem<Unit>?>(null)
 
     // cached transaction details (observable for Compose)
     val transactionDetailsCache: SnapshotStateMap<TxId, TransactionDetails> = mutableStateMapOf()
@@ -395,7 +392,7 @@ class WalletManager :
             }
 
             is WalletManagerReconcileMessage.PayjoinTxBroadcast -> {
-                payjoinTxBroadcast = true
+                payjoinTxBroadcast = TaggedItem(Unit)
             }
         }
     }
