+++ ios/Cove/WalletManager.swift
@@ -40,8 +40,9 @@ private extension WalletScanStatus {
     /// errors in SendFlow
     var sendFlowErrorAlert: TaggedItem<SendFlowErrorAlert>? = nil
 
-    /// set to true when a payjoin transaction has been broadcast (success or fallback)
-    var payjoinTxBroadcast: Bool = false
+    /// non-nil when a payjoin transaction has been broadcast (success or fallback);
+    /// UUID changes each time so onChange always fires even across multiple sends
+    var payjoinTxBroadcast: UUID? = nil
 
     /// cached transaction details
     var transactionDetails: [TxId: TransactionDetails] = [:]
@@ -304,7 +305,7 @@ private extension WalletScanStatus {
             }
 
         case .payjoinTxBroadcast:
-            self.payjoinTxBroadcast = true
+            self.payjoinTxBroadcast = UUID()
         }
     }
